Delete README.md
In this project, I applied my newly learned knowledge in statistical correlation and R programming skills. I analyzed a recent survey of college students. I analyzed if there is a significant relationship between the amount of time spent reading and watching television. As simple as it sounds, there are still many other factors that are taken into consideration. Also, other significant relationships were discovered in this analysis; the survey data is located in this StudentSurvey.csv file.